requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
coverage==7.8.2
cryptography==45.0.3
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==3.0.0
django==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-extensions==4.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-social-oauth2==3.1.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
factory-boy==3.3.3
faker==37.3.0
filelock==3.18.0
-e file:///home/iamprecieee/Desktop/HP-Backend
identify==2.6.12
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
loguru==0.7.3
markupsafe==3.0.2
nodeenv==1.9.1
oauthlib==3.2.2
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==4.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pygments==2.19.1
pyjwt==2.9.0
pyopenssl==25.1.0
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.11.1
python3-openid==3.2.0
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uuid6==2024.7.10
virtualenv==20.31.2
werkzeug==3.1.3
whitenoise==6.9.0
